adding coordinator protocol and clean up app delegate root
