+++ swift_algorithms/Coordinators/AppCoordinator.swift
@@ -1,19 +1,14 @@
 import UIKit
 
-final class AppCoordinator {
-    
-    private var rootTabCoordinator: RootTabCoordinator?
-    
-    var root: UIViewController?
-    
-    func rootViewController() -> UIViewController? {
-        let coordinator = RootTabCoordinator()
-        rootTabCoordinator = coordinator
-        root = coordinator.root
-        return root
-    }
+final class AppCoordinator: Coordinating {
+
+    let root: UIViewController?
+    var tabCoordinator: RootTabCoordinator?
     
     init() {
+        let coordinator = RootTabCoordinator()
+        tabCoordinator = coordinator
+        root = coordinator.root
         setAppearance()
     }
